Add unique name for link for accessibility
In cases where there are more than one Add to Calendar button per page, the various links need to be unique in order for users of assistive technology to be able to differentiate them.

Accessibility-wise, a link that simply says 'Google Calendar' isn't descriptive enough for someone using a screen reader to be able to use it to navigate successfullly.

This is my first time contributing to an open source thing, so please let me know if I'm doing this all wrong.