Implement 'All Day' option + update Yahoo
Yahoo now has an end time `et` parameter which allows it to be handled in a manner more consistent with ics and Google Calendar.

calculateEndTime() no longer calls formatTime - it only calculates the end time This was necessary to prevent functionality duplication in the yahoo implementation.

The example html file has also been updated.